+++ feature/login/sign-in/build.gradle.kts
@@ -5,6 +5,8 @@ plugins {
     id("lift.android.kotlin")
     id("lift.android.hilt")
     id("lift.android.test")
+    id("lift.android.oauth")
+    id("lift.android.google")
 }
 
 android {
