+++ feature/login/sign-in/src/main/kotlin/com/gradation/lift/feature/login/signIn/data/state/SignInScreenState.kt
@@ -4,24 +4,20 @@ import android.content.Context
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.remember
-import androidx.compose.ui.focus.FocusManager
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalFocusManager
 
 @Composable
 fun rememberSignInScreenState(
     snackbarHostState: SnackbarHostState = SnackbarHostState(),
-    focusManager: FocusManager = LocalFocusManager.current,
     context: Context = LocalContext.current,
 ): SignInScreenState =
-    remember(snackbarHostState, focusManager, context) {
-        SignInScreenState(snackbarHostState, focusManager, context)
+    remember(snackbarHostState, context) {
+        SignInScreenState(snackbarHostState, context)
     }
 
 
 data class SignInScreenState(
     val snackbarHostState: SnackbarHostState,
-    val focusManager: FocusManager,
     val context: Context,
 )
